+++ main.cpp
@@ -5,10 +5,12 @@
 #include <iostream>
 #include <vector>
 
-#include "xgboost/xgboost.h"
-#include "secureboost/secureboost.h"
-#include "../../defense/paillier/src/paillier.h"
-#include "../../defense/paillier/src/keygenerator.h"
+#include "src/aijack/defense/dp/core//rdp.cpp"
+#include "src/aijack/defense/dp/core//search.cpp"
+#include "src/aijack/defense/paillier/src/paillier.h"
+#include "src/aijack/defense/paillier/src/keygenerator.h"
+#include "src/aijack/collaborative/tree/xgboost/xgboost.h"
+#include "src/aijack/collaborative/tree/secureboost/secureboost.h"
 
 #define STRINGIFY(x) #x
 #define MACRO_STRINGIFY(x) STRINGIFY(x)
@@ -19,11 +21,40 @@ namespace py = pybind11;
 template <typename... Args>
 using overload_cast_ = pybind11::detail::overload_cast_impl<Args...>;
 
-PYBIND11_MODULE(aijack_secureboost, m)
+PYBIND11_MODULE(aijack_cpp_core, m)
 {
     m.doc() = R"pbdoc(
-        core of XGBoost
+        core of diferential_privacy
     )pbdoc";
+
+    m.def("eps_gaussian",
+          &eps_gaussian, R"pbdoc(eps_gaussian)pbdoc");
+
+    m.def("eps_laplace",
+          &eps_laplace, R"pbdoc(eps_laplace)pbdoc");
+
+    m.def("eps_randresp",
+          &eps_randresp, R"pbdoc(eps_randresp)pbdoc");
+
+    m.def("culc_tightupperbound_lowerbound_of_rdp_with_theorem6and8_of_zhu_2019",
+          &culc_tightupperbound_lowerbound_of_rdp_with_theorem6and8_of_zhu_2019,
+          R"pbdoc(culc_tightupperbound_lowerbound_of_rdp_with_theorem6and8_of_zhu_2019)pbdoc");
+
+    m.def("culc_upperbound_of_rdp_with_Sampled_Gaussian_Mechanism",
+          &culc_upperbound_of_rdp_with_Sampled_Gaussian_Mechanism,
+          R"pbdoc(culc_upperbound_of_rdp_with_Sampled_Gaussian_Mechanism)pbdoc");
+
+    m.def("_ternary_search",
+          &_ternary_search, R"pbdoc(_ternary_search)pbdoc");
+
+    m.def("_ternary_search_int",
+          &_ternary_search_int, R"pbdoc(_ternary_search_int)pbdoc");
+
+    m.def("_greedy_search",
+          &_greedy_search, R"pbdoc(_greedy_search)pbdoc");
+
+    m.def("_greedy_search_frac",
+          &_greedy_search_frac, R"pbdoc(_greey_search_frac)pbdoc");
 
     py::class_<PaillierKeyGenerator>(m, "PaillierKeyGenerator")
         .def(py::init<int>())
