+++ src/aijack/defense/dp/manager/accountant.py
@@ -1,6 +1,5 @@
 import numpy as np
-
-from aijack_dp_core import (
+from aijack_cpp_core import (
     _greedy_search,
     _greedy_search_frac,
     _ternary_search,
